requirements.txt
aiogram==3.10.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
